requirements.txt
python-dotenv==1.1.1
openai==1.99.6
fastapi==0.116.0
PyMuPDF==1.26.3
pydantic~=2.11.7
black==25.1.0
flake8==7.3.0
isort==6.0.1
uvicorn==0.35.0
python-multipart==0.0.20